# app/Http/Controllers/Api/ReserveController.php
<?php

declare(strict_types=1);

namespace App\Http\Controllers\Api;

use App\Billing\Enforcement\Contracts\ReservationStore;
use App\Billing\Enforcement\Upgrade\UpgradeGate;
use Cbox\Billing\Metering\Contracts\Enforcement;
use Cbox\Billing\Metering\Enums\DenialReason;
use Cbox\Billing\Metering\Enums\OutcomeStatus;
use Cbox\Billing\Metering\ValueObjects\BucketRequest;
use Cbox\Billing\Metering\ValueObjects\EnforcementOutcome;
use Cbox\Billing\Metering\ValueObjects\InfraFault;
use Illuminate\Contracts\Config\Repository as Config;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;

/**
 * `POST /api/v1/reserve` — reserve a set of meter buckets in one all-or-nothing call via
 * the engine's {@see Enforcement::reserveBucketsOutcome()}. The engine's three-way outcome
 * is mapped straight onto the JSON the SDK expects:
 *
 *  - `Allowed`       → `{outcome: "allowed", reservation_id}` (the held set is persisted
 *                       for the matching `/commit`).
 *  - `Denied`        → `{outcome: "denied", reason}` — a semantic refusal (unknown/disabled
 *                       meter, exhausted allowance/quota); fail-closed.
 *  - `Indeterminate` → `{outcome: "indeterminate", reason}` — a dependency was down; the
 *                       deployment's infra policy decided admit/refuse and it is signalled.
 */
class ReserveController extends ApiController
{
    public function __invoke(Request $request, Enforcement $enforcement, ReservationStore $reservations, Config $config, UpgradeGate $upgrades): JsonResponse
    {
        $request->validate([
            'org' => ['required', 'string'],
            'meters' => ['required', 'array', 'min:1'],
            'meters.*.meter' => ['required', 'string'],
            'meters.*.estimate' => ['required', 'integer', 'min:1'],
        ]);

        $org = $request->string('org')->toString();

        if ($denied = $this->denyUnlessMayActFor($request, $org)) {
            return $denied;
        }

        $requests = [];

        foreach ($request->array('meters') as $index => $meter) {
            $row = is_array($meter) ? $meter : [];

            $requests[] = new BucketRequest(
                $this->requireString($row['meter'] ?? null, sprintf('meters.%s.meter', $index)),
                $this->requireInt($row['estimate'] ?? null, sprintf('meters.%s.estimate', $index)),
            );
        }

        $outcome = $enforcement->reserveBucketsOutcome($org, $requests);

        if ($outcome->status === OutcomeStatus::Allowed) {
            return $this->allowed($outcome, $reservations, $config);
        }

        if ($outcome->status === OutcomeStatus::Denied) {
            $reason = $outcome->reason;

            $body = [
                'outcome' => 'denied',
                'reason' => $reason instanceof DenialReason ? $reason->value : 'denied',
            ];

            // The enforce→upgrade bridge (#52): a semantic denial carries the path to
            // unlock — the minimum reachable plan that grants the blocking meter and a
            // pre-built checkout deep-link to it. Omitted when no upgrade path exists
            // (already on the top plan, or nothing grants it).
            $upgrade = $upgrades->forReservation($org, $requests);

            if ($upgrade !== null) {
                $body['upgrade'] = $upgrade;
            }

            return new JsonResponse($body);
        }

        $fault = $outcome->fault;

        return new JsonResponse([
            'outcome' => 'indeterminate',
            'reason' => $fault instanceof InfraFault ? $fault->reason : 'dependency unavailable',
        ]);
    }

    private function allowed(EnforcementOutcome $outcome, ReservationStore $reservations, Config $config): JsonResponse
    {
        $set = $outcome->reservationSet();

        $ttl = $config->get('billing.api.reservation_ttl_seconds', 300);
        $ttl = is_numeric($ttl) ? (int) $ttl : 300;

        $reservations->put($set, $ttl);

        return new JsonResponse([
            'outcome' => 'allowed',
            'reservation_id' => $set->id,
        ]);
    }
}