# docs/openapi/cbox-billing.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cbox Billing API",
        "version": "1.0.0",
        "summary": "Enforcement + management HTTP API for the Cbox Billing application.",
...
        "contact": {
            "name": "Cbox Billing",
            "url": "https://github.com/cboxdk/cbox-billing"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/cboxdk/cbox-billing/blob/main/LICENSE"
        }
    },
    "servers": [
        {
            "url": "https://billing.example.com/api/v1",
            "description": "Your Cbox Billing deployment (replace host)."
        },
        {
            "url": "http://localhost/api/v1",
            "description": "Local development."
        }
    ],
...
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/reserve": {
            "post": {
                "operationId": "reserveBuckets",
                "tags": [
                    "Enforcement"
                ],
                "summary": "Reserve meter buckets (all-or-nothing)",
                "description": "Reserve a set of meter buckets in one all-or-nothing call. The engine's three-way\noutcome is returned verbatim:\n\n- `allowed` — the held set is persisted for the matching `/commit` and its\n  `reservation_id` is returned.\n- `denied` — a semantic refusal (unknown/disabled meter, exhausted allowance/quota).\n  Fail-closed. Carries an `upgrade` offer when a reachable plan would unblock it.\n- `indeterminate` — a dependency was down; the deployment's infra policy decided\n  admit/refuse and the reason is surfaced.\n\nNote: the HTTP status is always `200` — the decision is in the `outcome` field.\n",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReserveRequest"
                            },
                            "examples": {
                                "default": {
                                    "value": {
                                        "org": "org_acme",
                                        "meters": [
                                            {
                                                "meter": "api_calls",
                                                "estimate": 10
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The reservation outcome (allowed / denied / indeterminate).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReserveOutcome"
                                },
                                "examples": {
                                    "allowed": {
                                        "summary": "Allowed",
                                        "value": {
                                            "outcome": "allowed",
                                            "reservation_id": "rsv_7hKp2Qm9"
                                        }
                                    },
                                    "denied": {
                                        "summary": "Denied with an upgrade path",
                                        "value": {
                                            "outcome": "denied",
                                            "reason": "allowance_exhausted",
                                            "upgrade": {
                                                "required_plan": "pro",
                                                "checkout_url": "https://billing.example.com/billing/checkout/tok_abc123"
                                            }
                                        }
                                    },
                                    "indeterminate": {
                                        "summary": "Dependency unavailable",
                                        "value": {
                                            "outcome": "indeterminate",
                                            "reason": "dependency unavailable"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
        "/commit": {
            "post": {
                "operationId": "commitReservation",
                "tags": [
                    "Enforcement"
                ],
                "summary": "Commit a reservation to actual usage",
                "description": "Settle a held reservation to the actual usage per meter. Returns unused\nallowance/lease and appends one durable usage event per bucket with non-zero usage,\nthen drops the reservation. A missing/expired reservation is `404`; an actual that\nexceeds its reserved estimate is `422` from the engine's own validation.\n",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommitRequest"
                            },
                            "examples": {
                                "default": {
                                    "value": {
                                        "reservation_id": "rsv_7hKp2Qm9",
                                        "actuals": [
                                            {
                                                "meter": "api_calls",
                                                "actual": 8
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The reservation was committed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ok"
                                },
                                "examples": {
                                    "default": {
                                        "value": {
                                            "ok": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "description": "Unknown or expired reservation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "default": {
                                        "value": {
                                            "error": "Unknown or expired reservation."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "A Cbox Billing API token. The plaintext prefix encodes the billing plane: `cbl_`\n(live) or `cbt_` (test/sandbox). Operator tokens may act for any org; org-scoped\ntokens only for their own. Provision one with\n`php artisan billing:token \"<name>\" --org=<org-id>`.\n"
            }
        },
...
        "headers": {
            "XRateLimitLimit": {
                "description": "The per-window request ceiling for this token.",
                "schema": {
                    "type": "integer",
                    "example": 600
                }
            },
            "XRateLimitRemaining": {
                "description": "Requests remaining in the current window.",
                "schema": {
                    "type": "integer",
                    "example": 599
                }
            },
            "RetryAfter": {
                "description": "Seconds to wait before retrying (present on 429).",
                "schema": {
                    "type": "integer",
                    "example": 42
                }
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "No or invalid bearer token.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "missing": {
                                "value": {
                                    "error": "Missing bearer token."
                                }
                            },
                            "invalid": {
                                "value": {
                                    "error": "Invalid API token."
                                }
                            }
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "The token may not act for the requested organization (or lacks operator rights).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "default": {
                                "value": {
                                    "error": "This token is not permitted to act for the requested organization."
                                }
                            }
                        }
                    }
                }
            },
...
            "UnprocessableEntity": {
                "description": "Validation failed, or a business rule refused the request. Field validation returns\nLaravel's `{message, errors}` shape; a business refusal (invalid coupon, plan not\npriced in the currency, non-licensable plan) returns `{error}`.\n",
                "content": {
                    "application/json": {
                        "schema": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                {
                                    "$ref": "#/components/schemas/Error"
                                }
                            ]
                        },
                        "examples": {
                            "validation": {
                                "summary": "Field validation",
                                "value": {
                                    "message": "The plan field is required.",
                                    "errors": {
                                        "plan": [
                                            "The plan field is required."
                                        ]
                                    }
                                }
                            },
                            "business": {
                                "summary": "Business refusal",
                                "value": {
                                    "error": "Plan [pro] is not priced in EUR."
                                }
                            }
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "The per-token rate limit was exceeded.",
                "headers": {
                    "Retry-After": {
                        "$ref": "#/components/headers/RetryAfter"
                    },
                    "X-RateLimit-Limit": {
                        "$ref": "#/components/headers/XRateLimitLimit"
                    },
                    "X-RateLimit-Remaining": {
                        "$ref": "#/components/headers/XRateLimitRemaining"
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/RateLimitError"
                        },
                        "examples": {
                            "default": {
                                "value": {
                                    "message": "Too Many Attempts."
                                }
                            }
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "Error": {
                "type": "object",
                "description": "The app's canonical error shape for auth, scope, not-found, conflict, and business refusals.",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "A human-readable message."
                    }
                }
            },
            "ValidationError": {
                "type": "object",
                "description": "Laravel's field-validation error shape (thrown by request validation).",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "RateLimitError": {
                "type": "object",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Too Many Attempts."
                    }
                }
            },
            "Ok": {
                "type": "object",
                "required": [
                    "ok"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "const": true
                    }
                }
            },
...
            "ReserveRequest": {
                "type": "object",
                "required": [
                    "org",
                    "meters"
                ],
                "properties": {
                    "org": {
                        "type": "string"
                    },
                    "meters": {
                        "type": "array",
                        "minItems": 1,
                        "items": {
                            "type": "object",
                            "required": [
                                "meter",
                                "estimate"
                            ],
                            "properties": {
                                "meter": {
                                    "type": "string"
                                },
                                "estimate": {
                                    "type": "integer",
                                    "minimum": 1
                                }
                            }
                        }
                    }
                }
            },
            "ReserveOutcome": {
                "type": "object",
                "description": "The three-way reservation outcome. `outcome` is always present; the remaining\nfields depend on it (`reservation_id` when allowed; `reason` + optional `upgrade`\nwhen denied; `reason` when indeterminate).\n",
                "required": [
                    "outcome"
                ],
                "properties": {
                    "outcome": {
                        "type": "string",
                        "enum": [
                            "allowed",
                            "denied",
                            "indeterminate"
                        ]
                    },
                    "reservation_id": {
                        "type": "string",
                        "description": "Present when allowed — pass it to `/commit`."
                    },
                    "reason": {
                        "type": "string",
                        "description": "Present when denied or indeterminate."
                    },
                    "upgrade": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/UpgradeOffer"
                            }
                        ],
                        "description": "Present on a denial that a reachable plan would unblock."
                    }
                }
            },
            "CommitRequest": {
                "type": "object",
                "required": [
                    "reservation_id",
                    "actuals"
                ],
                "properties": {
                    "reservation_id": {
                        "type": "string"
                    },
                    "actuals": {
                        "type": "array",
                        "minItems": 1,
                        "items": {
                            "type": "object",
                            "required": [
                                "meter",
                                "actual"
                            ],
                            "properties": {
                                "meter": {
                                    "type": "string"
                                },
                                "actual": {
                                    "type": "integer",
                                    "minimum": 0
                                }
                            }
                        }
                    }
                }
            },
            "UpgradeOffer": {
                "type": "object",
                "description": "The path to unlock a blocked meter — the minimum reachable plan and a pre-built checkout deep-link.",
                "required": [
                    "required_plan"
                ],
                "properties": {
                    "required_plan": {
                        "type": "string"
                    },
                    "checkout_url": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "uri"
                    }
                }
            },
...
        }
    }
}